Don't apply enterprise state for oss builds.
