+++ dgraph/cmd/zero/license.go
@@ -1,0 +1,34 @@
+// +build oss
+
+/*
+ * Copyright 2018 Dgraph Labs, Inc. and Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package zero
+
+// dummy function as enterprise features are not available in oss binary.
+func (n *node) proposeTrialLicense() {
+	return
+}
+
+// dummy function as enterprise features are not available in oss binary.
+func (s *Server) updateEnterpriseState() {
+	return
+}
+
+// dummy function as enterprise features are not available in oss binary.
+func (s *Server) licenseExpiryWarning() {
+	return
+}
